moved lin/log space to utils
